Add files via upload
program to read an input text file in C and processing it accordingly, capitalizing specific words in a sentence read. There are some words that don't need to be capitalize so ignoring those words in a sentence when reading text file line by line. There can only be one capitalized word per line.